server/realtime/requirements.txt
pymercure
RPi.GPIO==0.7.0
gpiozero===1.5.1
python-dotenv==0.10.1
